Исправления №2. Исправлены заглавные буквы в названиях пакетов. Переписаны основные комментарии. Добавлены пустые строки в README.md.
