+++ pkg/day1/README.md
@@ -1,1 +1,2 @@
 В данном проекте содержится файл с решением задачи (https://leetcode.com/problems/two-sum/), а также тест к этому решению.
+
