+++ pkg/question/README.md
@@ -1,4 +1,6 @@
-В данной директории представлено решение задачи второго дня(https://leetcode.com/problems/isomorphic-strings/).
+В данной директории представлено решение задачи второго дня (https://leetcode.com/problems/isomorphic-strings/).
+
 Необходимо было написать функцию, принимающую в качестве аргументов 2 строки и возвращающую булевское значение,
 являются строки изоморфными или нет.
+
 Ещё содержится набор тестов для данной функции.
