+++ pkg/set/set_test.go
@@ -1,4 +1,4 @@
-package Set
+package set
 
 import (
 	"reflect"
