+++ pkg/worker/worker_test.go
@@ -1,4 +1,4 @@
-package Worker
+package worker
 
 import "testing"
